+++ CHANGELOG.rst
@@ -3,11 +3,13 @@ Changes
 
 5.0.1
 ~~~~~
-* FIX: ref-count leaks #372
-* ENH: Add support for building ABI3 wheels
+* ENH: Add %%lprun_all for more beginner-friendly profiling in IPython/Jupyter #383
 * FIX: mitigate speed regressions introduced in 5.0.0
 * ENH: Added capability to combine profiling data both programmatically (``LineStats.__add__()``) and via the CLI (``python -m line_profiler``) (#380, originally proposed in #219)
 * FIX: search function in online documentation
+* FIX: ref-count leaks #372
+* FIX: mitigate speed regressions introduced in 5.0.0 #376
+* FIX: Use import system to locate module file run by ``kernprof -m`` #389
 
 5.0.0
 ~~~~~
@@ -56,6 +58,7 @@ Changes
   * ``kernprof`` and ``python -m line_profiler`` CLI options
   * ``GlobalProfiler`` configurations, and
   * profiler output (e.g. ``LineProfiler.print_stats()``) formatting
+* ENH: Added capability to combine profiling data both programmatically (``LineStats.__add__()``) and via the CLI (``python -m line_profiler``) (#380, originally proposed in #219)
 
 4.2.0
 ~~~~~